build: the tree compiles without warnings under GCC 13
GCC 13 on the hosted runner warned about a possibly dangling reference where the manifest helpers took their key as a temporary string, two clamps written two to a line, a signed index compared with a block size, and the tests' malloc-backed operator delete, which it takes for a mismatch.

The helpers take literal keys, the clamps and the index are rewritten with the same results, and the tests' allocators say why that diagnostic is off around them.

+++ src/backends/cpu/cpu_backend.hpp
@@ -1751,7 +1751,7 @@ private:
             const uint8_t* y = row + b * gguf::Q8_0_TYPESIZE;
             uint16_t d16 = (uint16_t)(y[0] | ((uint16_t)y[1] << 8));
             float d = f16_to_f32(d16);
-            for (int j = 0; j < gguf::Q8_0_BLOCK; j++)
+            for (size_t j = 0; j < gguf::Q8_0_BLOCK; j++)
                 acc += (float)(int8_t)y[2 + j] * d * x[b * gguf::Q8_0_BLOCK + j];
         }
         return acc;
